+++ src/virtual-machine/src/main/java/org/smoothbuild/virtualmachine/bytecode/expr/base/BOperation.java
@@ -16,6 +16,7 @@ public abstract sealed class BOperation extends BExpr
     permits BCall,
         BChoose,
         BCombine,
+        BFold,
         BIf,
         BInvoke,
         BMap,