+++ src/virtual-machine/src/main/java/org/smoothbuild/virtualmachine/bytecode/kind/base/BOperationKind.java
@@ -7,6 +7,7 @@ public abstract sealed class BOperationKind extends BKind
     permits BCallKind,
         BChooseKind,
         BCombineKind,
+        BFoldKind,
         BIfKind,
         BInvokeKind,
         BMapKind,