+++ src/virtual-machine/src/main/java/org/smoothbuild/virtualmachine/bytecode/kind/base/KindId.java
@@ -20,6 +20,7 @@ public enum KindId {
   CHOOSE,
   CHOICE,
   SWITCH,
+  FOLD,
   ;
 
   public byte byteMarker() {